pyproject.toml
[project]
name = "shopify-graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "ollama>=0.4.8",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "urllib3==1.26.15",
    "whoosh>=2.7.4",
]
